- Add raw_timeformat parameter - PullRequest decline - User placed in common - DefaultReviewer inherits from User - Use get_time() Method for date values
