+++ atlassian/bitbucket/base.py
@@ -15,6 +15,7 @@ class BitbucketBase(AtlassianRestAPI):
 
         :return: nothing
         """
+        self.raw_timeformat = kwargs.pop("raw_timeformat")
         super(BitbucketBase, self).__init__(url, *args, **kwargs)
 
     def _get_paged(self, url, params=None, data=None, flags=None, trailing=None, absolute=False):
@@ -69,4 +70,5 @@ class BitbucketBase(AtlassianRestAPI):
             cloud=self.cloud,
             api_root=self.api_root,
             api_version=self.api_version,
+            raw_timeformat=self.raw_timeformat,
         )
