+++ atlassian/bitbucket/cloud/common/users.py
@@ -1,0 +1,22 @@
+from ..base import BitbucketCloudBase
+
+
+class User(BitbucketCloudBase):
+    def __init__(self, url, data, *args, **kwargs):
+        super(User, self).__init__(url, *args, data=data, expected_type="user", **kwargs)
+
+    @property
+    def display_name(self):
+        return str(self.get_data("display_name"))
+
+    @property
+    def nickname(self):
+        return self.get_data("nickname")
+
+    @property
+    def account_id(self):
+        return self.get_data("account_id")
+
+    @property
+    def uuid(self):
+        return self.get_data("uuid")
